+++ supercycleEngineApp/src/devs.dbd
@@ -1,3 +1,7 @@
 device(stringout, CONSTANT, devStrOutCtrl, "ControlStrOut")
 registrar(cmdMapStrOut)
 function(ioASubBuf)
+
+# from devObjString.cpp
+device(stringin, INST_IO, devStrInFromStr, "ObjPropString")
+#device(stringout, INST_IO, devSOFromString, "Obj Prop string")
